+++ src/router.tsx
@@ -1,6 +1,6 @@
 import App from 'App';
-import Home from 'pages/Home';
 import CreditCardRegister from 'pages/CreditCardRegister';
+import Home from 'pages/Home';
 import { createBrowserRouter } from 'react-router-dom';
 
 const router = createBrowserRouter([
